Attempt to normalize personal_sign data to a UTF-8 string
Until now the only attempted interpretation of personal_sign data was
to see if it was a sign-in with Ethereum message. In cases where the
provider call to sign was passed an encoded hex version of the UTF-8
sign-in with Ethereum message, the interpretation would fail, as the
`siwe` library expects to receive the string form rather than the hex
form of the sign-in message.

To fix this, and to display a user-friendly text message whenever
possible, the signing data parsing phase now checks for a hex string
and, in certain cases, attempts to treat is as an encoded UTF-8 string.

In particular, hex strings that are longer than 32 bytes and can be
parsed to UTF-8 and then re-serialized to hex without changes are
treated as valid strings and parsed as such before attempting any
further recognition.

`parseSigningData` returns the normalized UTF-8 string where applicable
if no further recognition succeeds.