Add an optional CoinGecko platform ID on Network
Because we have networks with ETH as a base asset (Arbitrum, Optimism)
and we don't want to fork ETH as a base asset throughout the codebase,
we instead optionally distinguish the CoinGecko platform ID at the
network level. This means Opti/Arbi ETH will always be 1:1 with base
mainnet ETH... which I believe is good wallet behavior for a rollup.