wrong-ledger-connected state includes required address
This information was formerly pulled from an AccountTotal object in
SignTransactionWrongLedgerConnected, but there is no reason for that
component to require the full account. Instead, `useSigningLedgerState`
now takes the address that is being requested and, when it determines
that a ledger is connected and isn't the correct ledger, that address is
included in the wrong-ledger-connected state as `requiredAddress`.

The SignTransactionWrongLedgerConnected component now takes the expected
address directly, as well. Additionally, that component uses
`truncateAddress` instead of doing a bespoke truncation.