default to transaction.gasPrice when receipt.effectiveGasPrice is not available.
Despite the ethers js docs stating that receipt.effectiveGasPrice "will simply be equal to the transaction gasPrice" on chains that do not support EIP-1559 - it seems that this is not yet the case with Optimism.

This commit fixes a bug where transaction enrichment was failing due to effectiveGasPrice being undefined and calling .toBigInt on it - but is not a perfect solution because transaction.gasPrice does not necessarily take into account L1 rollup fees.