Adjust alchemy tokenBalances response handling.
It seems that the api has changed to no longer have an `error` field on the response in cases where the response is valid.  Due to this change we were filtering out all valid token balance responses.  It is unclear if this api change is intentional or not so for now lets keep the b.error === null check.