Avoid adding transaction hashes for known asset transfers.
Now that we've moved to an approach of querying a given accounts most recent 1k asset transfers every 90 seconds - we need to ensure that we are not retrieving the same transactions unnecessarily.

Prior to this change - the "transactionsToRetrieve" queue would fill up every 90 seconds even after some transactions were correctly processed via "handleQueuedTransactionAlarm".  The performance hit from that was especially large on accounts with a large asset transfer history due to how we handle determining if a given transaction hash is already in the queue inside of "queueTransactionHashToRetrieve".