Replace newHeads subscription with a time-based block polling mechanism.
This change gives a number of improvements - including:
- Progress towards Dealchemification
- Significant reduction in background browser CPU usage
- Significant reduction in Alchemy CUP credit usage

Prior to this change - we were writing to disk and emitting an event that modified redux state at least once every ~2 seconds (Polygon block time).  This commit standardizes the rate at which we fetch latest blocks to once every 15 seconds, and moves us from a subscription-based method of fetching new blocks to a poll-based method.