Populate the `from` field when estimating transaction gas limit.
It is not uncommon for dapps to send us transactions without a gasLimit - which means we are responsible for populating it.  There are certain types of transactions (ERC-20 transferFrom for example) that will fail if the "from" field is not populated - which means that we were not able to estimate the gas limit for those transactions.