Block reject button when transaction is not ready
Same as sign button Reject button should be blocked
when transaction is not ready to be signed.
This is because sometimes there are pending dispatched actions when user
is rejecting transaction and there is no easy way to cancel them.
When they are finally dispatched then transaction should be already rejected but
it is being repopulated with data.