Update typing and behavior of chainService.db.getLatestBlock to handle case where a block might not be available.
It is reasonable to assume that either due to network conditions or an outage of a 3rd party api-provider our application might not recieve blocks for seconds, minutes, or hours.  While this will likely cause many other failures - this commit improves the functionality and typing of getLatestBlock to reflect that possibility.