Allow word wrap in TransactionDetailItem
For Ledger data display, some data displayed by the Ledger is long
enough that it will not display on one line correctly. Allow the browser
to break it up.

Ideally this would be broken up based on Ledger display character width
rather than browser wrapping, but that is left for later.