Move two i18n strings from signTransaction to ledger
These two referenced Ledger specifically, in a way that the main sign
transaction flow and namespace never should. The new signing flow will
properly scope references to these strings so that there's no need for
a double `useTranslation` call, but for the current
SignTransactionContainer an additional call is needed to pull in the
appropriate prefix for those two strings.