Push dismissedWarnings management into detail panel
This was being managed externally by the caller, but the state was
entirely internal to the component, and the external management was
solely to allow a different initial value. The initial value is now
accepted as a prop, while the state itself is managed inside
SignTransactionDetailPanel.

Notably, it's not clear the defaulting behavior should be happening in
the UI in the first place; instead the management being done that way
may need to be pushed back into the enrichment service. No action on
that front in this commit, however.