Add header exclusion to EIP*Info components
These are used to present signing data, and in the new signing flow the
headers will be provided by the container, so we add a flag that allows
the container to explicitly request that it control the data.

Also tweak a couple of style elements to more closely align to the
design.