Disallow bridge requests with TALLY_INTERNAL_ORIGIN
As a defense-in-depth measure, explicitly check for
TALLY_INTERNAL_ORIGIN being used over the provider bridge and disallow
it. This ensures the bridge will never forward internal-seeming calls
to the internal provider, and opens the door for the internal provider
to lend additional trust to internal calls without being concerned that
an external dApp can abuse them.