fix to-16 migration for green screen of death
An old migration didn't move the `ids` and `entities` from the address
level into the `chainID` level but copied it. Because of this these keys
are still in the state and there is is entities field inside the `ids` or
`entities` array. This has caused migration issues for some folks on
brave. The easy nondestructive fix is simply ignore these properties.
Because the slice will be refactored soon it did not worth the risk to
remove those keys from state.