Rework SignDataRequest as MessageSigningRequest discriminated union
SignDataRequest was an envelope container for (personal, non-typed-data)
message signing requests that were further subdivided by type, and
different message types (such as SIWE messages) had more specific
parsed data types in a `signingData` field. `SignDataMessageType` was an
integer enum that was used for message types.

The complicating factor with this approach was that trying to narrow
the types in consuming code was impossible, as the types were
represented without a way to type-guard the parsed message data. The
result was a need to check against `messageType` and the
`SignDataMessageType` enum, followed by an ostensibly unguarded `as`
cast for the parsed data in `signingData` itself.

The new approach uses a `MessageSigningRequest` envelope type that is
specialized by a type parameter, and that type parameter is a
discriminated union of possible signing data types called
`MessageSigningData`. The envelope type is structured the exact same
way as previously, but the typing allows for the `messageType` to be
used as a discriminant that guards the `signingData` type as well,
removing the need for an `as` cast and requiring instantiating code to
adhere to one of the discriminated type/data pairs rather than allowing
for instantiation of mixed types.

One other notable change here is moving away from enums, which are a
less favored TypeScript feature, to a simple discriminated union type of
string constants for the different message types.