Update getAccountTotal to use network correctly
`getAccountTotal` was using selectCurrentNetworkAccountTotalsByCategory,
which could lead to cases where the account total could not be returned
if the current network and the network of the account being requested
were not the same, even if there was data available for the network
being requested.

`getAccountTotal` is now rewritten in terms of a helper function,
`getNetworkAccountTotalsByCategory`, which takes the network as a
parameter instead of always having it fixed to the current network.
`selectCurrentNetworkAccountTotalsByCategory` is also rewritten to use
this helper, always passing the current network.