Signing top bar correctly sources network
The top bar had been adjusted to pull the network from the redux data
directly (transaction OR signing OR data signing), coupling it very
tightly to the details about how these are processed. This was done to
fix an issue where the network being shown was not the network that the
signature was for, but rather the currently-selected network in the
wallet, which could be different in certain scenarios.

In fact, the data being passed from PersonalSign, SignData, and
SignTransaction was failing to use available information to correctly
resolve the request's AccountTotal network, and those components were
subject to a bug in AccountTotal lookup, thus passing bad data to the
top bar, which triggered the usage of redux data.

This commit fixes the data lookup in the calling components, allowing
the top bar to eliminate its redux dependency and pull all data from
the passed AccountTotal.