Make logger redaction less aggressive
When the logger matched on a hex prefix, it would redact all non-space
characters that followed it. Because in many instances logging gets hit
with unformatted serialized JSON, this could lead to redacting an entire
log line because of one hex value.

Additionally, the redaction regex was using some unnecessary grouping
operators, though in practice this wasn't a big deal, and adding a
trailing space to better match against spaces, which is adjusted to use
an `|$` operator to handle end-of-string.