Change networks to be inactive at extension initialization.
In order to avoid querying for assetTransfers (by far our most expensive RPC query) and gas prices on all networks at application startup - we initialize all networks as inactive and only query for transfers when a network becomes active (e.g. a user actually opens the extension).

The tradeoff for our gain in efficiency here is that a user will need to be active on a given network in order for those network's transfers to appear in the extension.

Since it is impossible to open the extension directly on the "Activity" tab - and since Activities are only something we care about in the wallet (dapps generally don't use this information) the tradeoff seems minimal.