Fix indexingService test assertion.
There was not much value to asserting the order in which assets were
being emitted (since that is not important to our application logic).  Due to this the first assertion has been removed and a comment has been added inside of the first event handler clarifying why we still want an empty event handler in the test.