Throttle handlePriceAlarm to every 5-seconds at most.
There are two common scenarios happening that may rate-limit coingecko for users (and result in inconsistent prices).

1.  When the extension is first loaded - chain service emits every account on every network as part of the `newAccountToTrack` event - each of which triggers a call to `handlePricealarm`.

2.  When a user switches to a network for the first time - a `newAccountToTrack` event is emitted for every account in the wallet.

By throttling the `handlePriceAlarm` function - we ensure that these bursts of `newAccounttoTrack` events do not result in repeated duplicate calls to coinGecko - and mitigate the risk of the user exceeding the coingecko API limits.