Support custom SharedAssetIcon sizes
In certain cases, a shared asset icon will need to be used that is not
one of the standard small/medium/large sizes. For these cases, the
`size` prop can now be set to a `number`, in which case the icon
width/height will be set to that size instead of one of the standard
values.