Properly collapse account name/address in top menu
When the top menu's account name/address runs out of space, it now
properly elides. To do this, declare that the connection status button
is inflexible, and give a min-width of 0 to the components in the
hierarchy that contain the account name and address so that they will
flex properly if space is insufficient.

This change applies to both the top menu in the regular wallet and on
the signing page.