Transition icon colors on hover where relevant
In a few places, icons were being used alongside text whose color
transitioned on hover. Because the icons are not colored the same way,
the icon colors would change instantly instead of transitioning, which
led to a jerkiness in the hover animation.

SharedIcon now has a transitionHoverTime property that sets the time it
takes to transition the icon color on hover. This is applied
appropriately for the account lock/unlock button in the accounts panel,
as well as for the address copy button on the Send page.