update nonce management to be eip-1559 aware
If a chain does not support EIP-1559/doesn't have mempool then
the tx is either picked up or dropped immediately. Eg on optimism
and arbitrum. This means that we should rely on the nonce that
we got from the chain and no need for us to track lasSeenNonce.

The same distinction should also happen when releasing nonces.

Note: a more robust nonce management which take into account
the nonces in the mempool and the fact that they might not be
linear will be a follow up work.