extend the transferAsset with nonce option
This addition will create the path for canceling transactions
and also user defined nonce handling.

For this to work the nonce removal had to be moved into
the provider bridge service. This should have been it's
place originally. The reason for this is that a dApp can't
know the correct nonce to assign to any transaction
that it sends to be signed. That's because the number
of sign requests is not a published information.