clean txs from our registry which are not on chain
In certain edge cases it can happen, that we have stored
transactions that did not get registered on chain. Before
this change we didn't have a way to clean those up.

This change implements a public `getOrCancelTransaction`
that can be used later to trigger the clean up from the UI.

These type of error should be also cleaned up automatically,
so this function is also called from `retrieveTransaction`.
This way if any faulty tx is registered will be cleaned up
on `queuedTransactions` alarm. (The same cleanup
would have happened after `TRANSACTION_CHECK_LIFETIME_MS`
but since that's 10h it was perceived as a bug by the users.