Inject environment blocks from PR descriptions in builds
A new JS script executed with actions/github-script checks the PR
description for a section with heading `Testing Env` or
`Testing Environment`. This section is expected to be immediately
followed by blank lines and then a code fenced-block with
`dotenv`-syntax information.

The information is further filtered to only allow ENABLE_, USE_, SHOW_,
and SUPPORT_-prefixed environment variables. Any deviation from this
will be logged to let the user know the script filtered the environment
down.

Finally, the filtered environment block is returned as the workflow
step's output. The dev build step emits this output into a `.env` file
for build usage.