add all tx to retrieve queue for health check
Originally listeners were created for transactions that are pending.
But if a transaction has wrong nonce it will be dropped by the mempool.
This takes ~5-10 minutes. In this case the listener will never fire
and the tx will be stuck.

In this solution we can use the `transactionsToRetrieve` to check
if the chain still knows about the tx, and if not we can mark dropped.

There was a problem with the existing queuing solution, because it
was not editable at runtime. This has caused the tx to be stuck until
extension reload.

A queueing solution should be put in place, but so far couldn't find
an implementation that didn't felt overly complex or hacky.