Fix issue with cut off keyring unlock button
To do this, the keyring step has its cancel button removed when in the
new signing flow---because the keyring unlock used to precede signing
confirmation, there needed to be a way to cancel early. Now that it is
presented after, this is less critical.

Additionally, the styling around footer handling for the new signing
flow is adjusted so that screens with no footer can take up the full
extension popup height.