Don't insert build links for non-PRs, throw for errors
Any API requests that fail should throw errors to fail the action,
rather than just logging an error and exiting successfully, otherwise
it will be hard to notice when something goes wrong.

To make this work correctly, builds triggered by workflows that are run
on push instead of PR need to exit successfully, since push builds are
for main merges that have no associated PR, so there is no description
to update.

The post build link script now detects cases where the artifact name
has a commit id rather than a PR number, and exits successfully with a
note that the workflow was for a merge commit and therefore no build
link will be generated. In other failure cases, it throws an error.