Fix post-build-link Error constructor calls
These were being called as if they were console logs, with multiple
parameters meant to be concatenated into one string. TypeScript does not
seem to flag this as an error, but nonetheless it leads all strings but
the first being dropped in the error message.

These are now handled as string interpolations.