Move from a recursive primary provider reconnect strategy to an iterative one.
Reconnecting to a primary provider that is not a websocket provider was resulting in a geometric explosion of reconnect attempts.  Since each call to attemptToReconnectToPrimaryProvider resulted in a recursive call to attemptToReconnectToPrimaryProvider on failed connection - as well as a call to attemptToReconnectToPrimaryProvider inside of resubscribe if the primary provider was not a WebsocketProvider - the reconnection logic quickly drained the browser of resources after just a few minutes of trying to reconnect.

By replacing the recursive strategy with an iterative one - we can get rid of the geometric increase and make the reconnect logic a little bit easier to reason about.