Remove reconnect primary provider logic from resubscribe.
No need for this any longer since we check for primary connection every 15 seconds now.  It seems that the "provider instanceof WebSocketProvider" check was a proxy for "Is this the primary provider" and since the primary provider might not necessarily be a websocket provider (anymore) this check is now irrelevant.