Push windows.create popup to same height as icon popup
Because windows.create, as used to create popups for signing, takes a
height that includes the native browser window frame, popups created
this way are slightly shorter than the popup triggered by the Tally
toolbar icon.

To fix this, popup.ts always runs a `resizeBy` call when it loads that
adjusts the window height by the window frame's size (approximated by
subtracting inner from outer height). This results in a brief jump in
the window size, but it's small enough (on macOS) that the eye is not
drawn to it.