WIP complete root element structure rework
To simplify scroll handling, clarify ownership of styles, remove render
warnings and glitches, and fix types in Popup.

- CorePage is now used by page components, not forcibly wrapped in the
  router.
- CorePage owns, through props, rendering or not rendering the top and
  tab bars.
- All transitions are removed.
- Proper typing is used in Popup.tsx, ts-expect-error usage is gone.
- Fixed positioning is largely removed for anything other than popovers
  and slideovers.
- Less z-indexing.
- Wallet asset list element structure more clearly reflects the quick
  links to swap and send as distinct, non-hierarchically-descendant of
  the asset item link.

Still missing are several invocations of CorePage, and an attempt to
reintroduce page transitions with reduced complexity.

Also the functionality that resets the tab selection in wallet activity
when the NFT availability changes is currently not working, but no such
networks exist by default. Some more work to be done there.