Add explicit reconnect logic for the alchemy provider.
Since our current reconnect logic in SerialFallbackProvider revolves around detecting if the primary provider is connected - and if not reconnecting to it - we need a way to handle an edge case where the connection of the alchemy websocket provider may drop despite the primary provider still being connected.  A simple solution here is to check its connection status every 15 seconds and reconnecting if the provider is disconnected.