Stop marking account activity when a dapp requests user accounts.
The "lastUserActivityOnNetwork" and "lastUserActivityOnAddress" properties are exclusively used to gate pull-based lookups for user assets / activites on networks.  markAccountActivity is responsible for updating these properties - but there is no utility in them being updated if the wallet is not open.  Since we already mark account activity via the "userActivityEncountered" and "newSelectedAccount" events when a user opens the wallet and selects a new account (respectively) there is no need for us to use a background action like connecting to a dapp to mark account activity.