Set a long timeout on transaction enrichment.
We don't want transaction enrichment to block transaction signing on users with poor network connections.  In particular - users with low internet speed and high latency can experience an effect similar to backpressure if a balance or balance refresh alarm is triggered between populating a tarnsaction request and enriching a transaction signature.

Due to the number of RPC and UNS requests involved in transaction enrichment a user with just 400ms of network latency might need to wait up to a minute to complete transaction enrichment (even though the size of the data transmitted is small) due to the serial nature in which we enrich transactions.