Limit RNS Name resolution to Rootstock addressnetwork lookups.
Having every single name service lookup a possible name for every address has become a bit too costly.  We need to start limiting name services to the networks that they address and this is a start to that.

Will hold off on limiting ENS's NameResolution logic because it is so ubiquitous that its not yet clear what networks should and should not use ENS name resolution.