Add backdrop filter hook
Chrome seems to have problems when elements with
backdrop style are rendered initially out of the
viewport - browser is not rendering them at all.
This is a try to create a workaround to force them to rerender.

Scrolling in and out of the view is still breaking it, needs more work.

Co-Authored-By: Jorge Luis <28708889+hyphenized@users.noreply.github.com>