Remove exponential backoff logic in SerialBackoffProvider retries.
Now that we support more than 1 rpc endpoint per network - our style of backoff should be breadth-first rather than depth-first.

That is to say - rather than trying one disconnected provider multiple times before switching to attemps on a new provider - we can and should quickly move on from providers that are disconnected and attempt to send messages on alternate providers.