refactor transaction retrieval scheduling
We need to distribute the transaction requests in time
otherwise it can cause a performance hit on the users's
machine and rate limitations with the providers.

But we need to be prepared for the background script shut
down any time so we need to be in sync with what's persisted.

The alarms can fire 1+ minute intervals, so we need an internal
more granular timer