Add a method to cleanup per-provider cache of eth_getBalance and eth_hasCode results.
While the cacheing of eth_getBalance and eth_hasCode significantly improves performance and reduces rpc usage on initial wallet load - its current implementation leaves room for the possibility for the cache to grow large enough as to cause the user to run out of memory.  A periodic check that both reports the size of the cache (for the purposes of debugging user issues) and reduces it should help mitigate this issue until we get to a point where we no longer need the cache.