Remove provider pendingTransaction subscriptions.
Our current method of pending transaction subscription does not seem to be providing much benefit while resulting in either:

a) Large infrastructure costs when subscribing via alchemy (their pricing model seems to charge us for every pending transaction on a given chain - not just the transactions that match the filter we specify)
b) Large CPU load when subscribing via a generic provider (we need to poll for every transaction in the mempool using its transaction hash and then serialize it).

The main benefit we get out of subscribing to pending transactions in the mempool is we are able to detect and display transactions that relevant to a given account even if those transactions were not initiated by the Tally Ho wallet (e.g. a user has their account imported to both tally ho and rainbow and make a transaction using rainbow).

It seems reasonable to assume that this scenario is both somewhat rare and already taken care of by our periodic polling with alchemy_getAssetTransfers (although in that case the transactions can take up to 1 minute to show up).

With the above points in mind and keeping in mind that we subscribe to transactions sent from the wallet via provider.once - we've chosen to remove our subscription to pending transactions until we find a use case that justifies the infra cost and/or cpu overhead generated by the subscription.