Use async api of local storage in logging.
In order to keep logger api sync, and avoid logs to overwrite each others, a queue is used to process the log entries asynchronously.

genericLogger() function has been split by functionality:
 - prepareLogEntry(): decorates, and saves additional data onto the log entry
 - writeToConsole(): writes the entry to the console immediately
 - saveLog(): saves the entry  into the local storage

This splitting is necessary, because these operations are not called in one turn anymore because of the async queue processing.

The purpose of the usage of isInitialized property is to prevent queue processing before the initialization completes since it is also async.
We don't have to wait for the results of the logger.init() call this way.