Improve asset merge logic.
Previously we were removing coinType when merging assets if the second asset merged did not have a coinType.  Additionally - we were removing tokenlists if the first asset merged did not have a tokenlist.  This commit addresses these issues and adds a regression test to ensure this does not happen again.