Separate wallet_addEthereumChain from wallet_switchEthereumChain.
Now that we're building in the ability to add custom networks we need to deliniate these two methods.  Lets keep wallet_switchEthereumChain functionality inside of wallet_addEthereumChain in case users try to add networks that are already supported - but give a separate path for the wallet_addEthereumChain method to interact with chainService.