Improve logger implementation
Removed dependency on webext browser polyfill. The logging context is now setup
at script initialization, per execution context. This results in one separate
instance for injected content-script, background script, and extension UI.
The context id, which is used as a prefix for logs is now set to "BG" for
the background script and "UI" for logging from extension pages.