Add base asset even if balance is zero
For accounts that are not in a "loading" state we were not adding
base assets if their balance was zero. Account could be past their loading
state if there was an ENS name loaded faster than balances.
That way some accounts has empty assets list but we should always
show at least base assets.