Fix CODEOWNERS teams and drop package.json
CODEOWNERS was still referencing tallyhowallet as the org
name for teams, which is incorrect. This is now correctly
tahowallet. While we're in here, add an extension-security-auditors
group to handle ownership of CODEOWNERS itself as well as
keyring code.

As a last adjustment, drop package.json, which was triggering CODEOWNERS
during releases because of a simple bump for the top-level package
version. Instead, yarn.lock remains under CODEOWNERS. Since
CI requires yarn.lock be up to date by using `--frozen-lockfile`,
the effect of requiring that dependency changes be audited by
the extension-dependency-auditors group remains.