Delete accounts for a given network when deleting that network.
Since we rely on fetching existing accounts to set up initial state - if we have accounts on corresponding to a network without the corresponding network - our state setup fails on application load.  To prevent this - we should delete accounts for a given network along with the rpcUrls, base assets, and networks when removing the network.