Prefer disable throttling when querying CoinGecko API.
This change was introduced to resolve an issue where 429 responses from CoinGecko would block swap functionality in the application.  This was happening because the fetchJson function we were using automatically waits and retries requests that come back with a 429 status code.  Due to the high likelihood of us being rate limited when hitting the coingecko api - this is actually undesirable functinality since we want to fail fast in the case that coingecko is unavailable.

By passing in a throttleCallback to fetchJson that always resolves to false  - we can rely on the (already existing) code around the function call to handle cases where we get a 429 or a bad response.  This results in a snappier user experience when querying prices - and does not break any existing functionality - all that is changing is that we now handle 429's how we would handle 400's and 500's - instead of special-casing them.