Run e2e tests also for PRs which modify feature flag
We want to run `e2e-tests` job not only after merges to `main` and for PRs that
have the `e2e` substring in the feature branch name, but also for PRs that
modify the feature flags. We do that by adding a new job
(`detect-if-flag-changed`) that checks if the '.env.defaults' file (which stores
the feature flags configuration) is modified by the PR. The job uses
https://github.com/marketplace/actions/paths-changes-filter action to produce a
true/false output concluding if the file was modified. If value is `true`, the
`e2e-tests` job will be executed (other criteria for execution of the job remain
unchanged.