Update link to the Discord channel for bug reporting
We've been linking the `#feedback` Discord channel on the New GH Issue landing
page. The channel no longer exists and the `#bug-report` channel is now a
designated place for bug reporting in Discord.
Updating name of the channel and the invite link. The invite link does not have
the expiry date and allows for unlimited use.