Update link to the Discord channel for bug reporting
We've been using link to the `#feedback` Discord channel in the Bug Report
template. The channel no longer exists and the `#bug-report` channel is now a
designated place for bug reporting in Discord.
Updating the invite link. The invite link does not have the expiry date and allows
for unlimited use.