Add workflow commenting release PRs with tests list
The commit introduces a GH ACtions workflow which adds a comment with tests list
to the PRs with the release candidates (PRs that want to merge `release-*`
branches to the `main` branch). The test list should be specified in the
`./.github/workflows/test-list/release-test-list.md` file. The comment will be
added only once, right after the PR gets created.