Add framework and E2E tests for testing transactions
This commit introduces automated E2E tests for asset transactions and changes to
the testing framework needed for the tests to run successfully.

Changes to the framework include
* Creation of a Hardhat project in `./ci` folder. In the project we configure a
`hardhat` network, which is a network that's a fork of Mainnet. The fork begins
on the block specified in the `blockNumber` parameter in `hardhat.config.ts`
(currently it's set to `17094516`, which is a starting point for our fork-based
E2E tests). Any actions performed on the `hardhat` network will only be applied
locally, they will not affect the real Mainnet network.
* Creation of the helper files and functions in `./e2e-tests/utils`.
* Modification of the GH ACtions workflow. A new job (`e2e-tests-fork`) is added
to the `build.yml` workflow. The job uses the build artifact from earlier jobs,
installs the project, installs Hardhat and runs the local `hardhat` network with
the Mainnet fork. Then it runs the Playwright E2E tests designed for fork (the
ones with `-fork` suffix) and upload the test artifacts in case of tests failure.
Also the old `e2e-tests` job gets modified - it's now running only those E2E
tests which are not intended for running on fork (don't contain the `-fork`
suffix.)

The following tests got added:
* User can send base asset (executed on forked Mainnet)
* User can open ERC-20 transfer from asset list and can reject the transfer
  (executed on forked Mainnet)
* User can send ERC-20 from header (executed on forked Mainnet)
* User can send base asset (executed on Goerli testnet)
Most of the tests are run on Mainnet fork (to not waste the funds), but one test
is run on Goerli - this is to verify the display of the transaction on the
activity views (activity history doesn't work currently on fork).