Categorize accounts by signer id instead of just keyring id
Previously, the only type of id being taken into account to
differentiate between different signers was the keyring id. This meant
that two accounts from two different Ledgers appeared under the same
entry in the account list, "Ledger 1". Additionally, it meant if you
renamed the Ledger category, it was for all Ledgers, not just for one.

The underlying change to fix this is fairly minor: keyringId is now
generalized to signerId, which is derived separately for each account
signer type. The derivation is set up such that adding a signer type
will require specifying how to differentiate between signers of that
type. Any signers that prefer not to differentiate can return `null`, or
another value.