Add a SigningService hook for pre-notifying signers of a request
Some signers can require various types of setup when a new signing
request is about to come in. Two quick examples are hardware wallets,
which may need to explicitly refresh the connection state to make sure
that the wallet is still connected and ready, or remote signers like
WalletConnect (currently not supported), which might need to establish a
remote connection to be able to make a request.

For the case where a Ledger did not have arbitrary data signing enabled
specifically, 8dc160b landed a fix where a full connection refresh would
be kicked off, fixing a specific bug. However, this led to an issue
where arbitrary data signing information would be checked in certain
cases when a Ledger was disconnected, throwing an error early in the
signing process and preventing advancing to the signing stage even for
non-Ledger signers.

This commit reworks that fix, instead providing a new method on the
SigningService, which is responsible for coordinating between signers.
This method, `prepareForSigningRequest`, is called immediately prior to
passing a signing request (data, typed data, or transaction signing) to
the UI, giving the signers an opportunity to do any setup they might
need.

The method can be used in blocking fashion in cases where the UI will be
presented quickly and the signer should already be prepared when the UI
is presented; this is used for transaction signing and data signing. For
typed data signing, the data is enriched before presenting the UI, and
this will generally take much longer than the current implementation of
signer preparation, so the call to `prepareForSigningRequest` is left
un-awaited.

Currently, the signing service's implementation simply always asks the
Ledger service to refresh its connection information in preparation for
the request. Future refinement could pass the signing address and check
whether the Ledger service is responsible for that address before
refreshing connection information, but the current implementation seems
to have pretty low overhead.